* commit '5a48acc58aff18805fc98dc79abdd77caafeaefa':
  CAL-479 prevent default calendar deletion

+++ lib/JSON/Plugin.php
@@ -163,6 +163,13 @@ class Plugin extends \Sabre\CalDAV\Plugin {
     }
 
     function deleteNode($nodePath, $node) {
+        if ($node->getName() === \ESN\CalDAV\Backend\Esn::EVENTS_URI) {
+            return [403, [
+                'status' => 403,
+                'message' => 'Forbidden: You can not delete your main calendar'
+            ]];
+        }
+
         $this->server->tree->delete($nodePath);
         return [204, null];
     }
